[Cargo] #cargo_query Fix issues with quotes and other parsing
CargoUtils.php
==============

(1) Fix smartSplit so that parenthesis, separators and "the other quote" (single quote in a double
quoted string or double quote in a single quoted string) inside a quoted string are not considered
lexically.

(2) Add 3 functions to CargoUtils to prepare regex expressions to match table names and field
names, allowing for '$' as part of the identifier.


CargoSQLQuery.php
=================

(3) Change the quote elimination logic in validateValues so that "the other quote" inside a quoted
string is not considered lexically. Earier double quotes were made into single quotes so if a
double quote happened inside a single quoted string (or the other way around) things would get out
of synch.

(4) Eliminate quoted strings before all checks within validateValues, but for $limitStr -- i.e.
$fieldsStr, $whereStr, $joinStr, $groupByStr, $havingStr and $orderByStr.

(5) Fix sqlFunctionRegex within getAndValidateSQLFunctions to allow for multiple white spaces '\s'
between the function name and the parenthesis.

(6) Change setDescriptionsForFields, handleVirtualFields and addTablePrefixes use more resilent
regex when mathing identifiers.

(7) Bug T120583: Change the HOLDS check to match the field "book", but not the field "bookworm"
when the field that allows for multiple values is "book". Makes the match case-insensitive since
table and column names are usually not case sensitive.

(8) Change the decomposition of table / field from explode('.') to a regex, adding resilience.

(9) Fix the string literal identification in setDescriptionsForFields to non-greedy.

(10) Add support for double quoted literal strings in setDescriptionsForFields.

(11) Fix an issue where quoted strings were being scanned for function calls throwing and
exception when an open parenthesis followed a word inside a quoted string.

Change-Id: I56e9bee6237e1bfd66889bf9a63b2b3216ebd2d3