Don't return empty strings from CargoUtils::smartSplit()
This was prompted by a requirement to ignore empty field
definitions in cargo_query: If a field list ended in a comma
(indicating an empty trailing field definition) then an
"Uninitialized string offset: 0" was being thrown. This change
avoids this by discarding empty field definitions before
processing them, via modifying the smartSplit function to
never return empty strings.

This also includes a couple of code-style fixes for array syntax,
and the start of some a test for the smartSplit function.

Change-Id: I7cf9160b3a7ee7d341745c8b2347221502f9c22e