Remove mention of wgMonthNames
Doesn't seem particulary relevant to this extension. It has
its own PHP code to export it, and own JS code to read it.

The var is being deprecated in core (T219340).

Update the comment to clarify why one would make it 1-index,
and clarify that that would no longer be an array in JSON, but
rather an object.

Change-Id: Ib35f25533cc80ce112df714d2223cac32b6803a3