Use of SQL subqueries for every HOLDS occurrence
HOLDS now uses individual subqueries for each clause which are connected using the given operators. The required IDs are extracted for further assessment.
This way it avoids the use of other SQL statements such as join and thus has no effect on these parameters if defined by the user.
This also enables the support to use  multiple HOLDS queries on the same field.

Change-Id: I30570c0882143b2805134e3570505143f2e7c7d5