Add config "wgCargoLegacyStoreNullAsEmptyString"
In I58f44498e66c98e8389459365c235583064075c6, Cargo was patched so that
missing values properly returned null instead of empty string. But this
can cause a lot of breaking changes especially for wikis that use Lua,
so this patch introduces a parameter to restore the legacy behavior, but
defaults it to false.

Change-Id: Iecf457e48040b680c1a5997faaee6ddc1c96c2eb