Add $wgCargoStoreUseTemplateArgsFallback setting
Additionally, we now fully check non-emptiness of fallback params
prior to adding any additional array keys. This is a better fix for
T290165 than I58f44498e66c98e8389459365c235583064075c6 in that it
never results in any type-mismatches in the database.

A future commit will properly implement the config
$wgCargoLegacyStoreNullAsEmptyString in CargoLuaLibrary.

Resolves T310389

Change-Id: Ib07b015f36569f41c678815e052be05fc8f4b454