Add wgCargoLegacyNullLuaFieldsAsEmptyString
This patch rverts Iecf457e48040b680c1a5997faaee6ddc1c96c2eb which
added a config param wgCargoLegacyStoreNullAsEmptyString. That param
addressed the issue on the storing end, but it's better to address on
the querying end directly.

It turns out the original backwards-compatibility break happened in
I1787cdebb0d9389697e0648193bba7e0d72e6fc9 and so this is a much more
localized change that affects only Lua queries and nothing else.

Change-Id: I70c50ff0fd8394c50b18f8f827be068186f1a24e