Fix PageSaveComplete hook handler
8039cb1897b754e13963b0e95a949a017ee6b842 updated Cargo's page move
handler to use the new PageSaveComplete hook. Unfortunately, the updated
logic tries to pass an undefined variable `$title` to
storeValuesForPage(), so it fails.

Phan would have caught this issue if CargoHooks.php was within
includes/, which is scanned by Phan. So, move CargoHooks.php to live
with the rest of the PHP code under includes/, and fix Phan issues
within this file. This required adding some stubs for AdminLinks
extension classes.

Change-Id: I9bf7e81a35f3d58c9f7382312de3f341553ed794