CargoQueryAPI: Expose field alias errors
`action=cargoquery` rejects queries that contain field aliases starting
with an underscore (_). However, if the wiki is configured to not
display exception details to everyone, this error is not visible to the
user that issued the API request. This can cause confusion.

As a fix, expose this error as a regular API error so that users can fix
their API queries accordingly.

Bug: T334755
Change-Id: Icd12baae36341b6c1d13435674a900758dbbcdd6