CargoStore: Fix false being treated as empty for Scribunto
CargoLuaLibrary::cargoStore() calls CargoStore::storeTable() with the
table passed from Lua without any pre-processing done on it. The table
is eventually passed to CargoStore::blankOrRejectBadData(), where a
check of `$fieldValue == ''` is done.

If Lua code passed a `false` for the value, this would evaluate to
`false == ''`, which is true thanks to type juggling. This patch does
a strict equality check for an empty string and null instead to ensure
that no empty values are passed for mandatory columns whilst properly
handling false.

However, the code will later run afoul of another bug, T370592, which
this patch does not address.

Bug: T370590
Change-Id: I541bede8515e9e4a0fd4999b2b9f36db81c837b8