CargoFieldDescription: Fix storing false from Scribunto
CargoLuaLibrary::cargoStore() calls CargoStore::storeTable() with the
table passed from Lua without any pre-processing done on it. The table
is eventually passed to CargoFieldDescription::prepareAndValidateValue(),
where `trim( $fieldValue )` is called.

If Lua passes false as a table value to mw.ext.cargo.store(),
`trim( false )` is called. This (and `(string)false`) returns an empty
string. This patch sidesteps this by using `'0'` if false is passed,
and `trim( $fieldValue )` otherwise.

Bug: T370590
Bug: T370592
Change-Id: Ia3dcd3247d6b939e385066ad925f5e391178957a